CaseStudy/requirements.txt
django ~=3.2.10
djangorestframework_simplejwt==4.5.0
djangorestframework == 3.14.0
django-rest-knox==4.2.0
PyJWT==1.7.1
django-cors-headers==3.7.0
django-filter==2.4.0

